Updated Newsfeed component to Homepage component
